Moves methods that touch the data from util to lib, updates "cl" handler to new style
