+++ e3sm_to_cmip/cmor_handlers/cli.py.bak
@@ -11,9 +11,9 @@ import logging
 from progressbar import progressbar
 
 from e3sm_to_cmip.util import print_message
-from e3sm_to_cmip.util import get_dimension_data
+from e3sm_to_cmip.lib import get_dimension_data
 from e3sm_to_cmip.util import setup_cmor
-from e3sm_to_cmip.util import load_axis
+from e3sm_to_cmip.lib import load_axis
 
 # list of raw variable names needed
 RAW_VARIABLES = [str('CLDICE')]
